Move code for adding an assignment to a separate page.
